requirement.txt
Adafruit-Blinka==8.24.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-neopixel==6.3.10
adafruit-circuitpython-pixelbuf==2.0.3
adafruit-circuitpython-requests==2.0.2
adafruit-circuitpython-typing==1.9.5
Adafruit-PlatformDetect==3.53.0
Adafruit-PureIO==1.1.11
Faker==12.0.1
pyftdi==0.54.0
pygame==2.5.2
pyserial==3.5
python-dateutil==2.8.2
pyusb==1.2.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
six==1.16.0
sysv-ipc==1.1.0
typing-extensions==4.7.1
